requirements.txt
blinker==1.4
click==6.6
facebook-sdk==2.0.0
Flask==0.11.1
Flask-API==0.6.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-Stormpath==0.4.8
Flask-WTF==0.13.1
gunicorn==19.6.0
httplib2==0.9.2
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2client==1.5.2
oauthlib==1.0.3
passlib==1.6.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyDispatcher==2.0.5
PyJWT==1.4.2
PyMySQL==0.7.9
python-dateutil==2.6.0
requests==2.12.1
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.4
stormpath==2.4.4
Werkzeug==0.11.11
WTForms==2.1
